+++ app/src/main/java/com/example/buyerapp/data/network/dto/StoreNotifyStateReq.kt
@@ -2,5 +2,5 @@ package com.example.buyerapp.data.network.dto
 
 data class StoreNotifyStateReq (
     val active: Boolean,
-    val storeId: Int
+    val storeId: Long
 )
