+++ app/src/main/java/com/example/buyerapp/domain/usecase/SaveStoreUseCase.kt
@@ -2,7 +2,6 @@ package com.example.buyerapp.domain.usecase
 
 import com.example.buyerapp.core.framework.usecase.LocalUseCase
 import com.example.buyerapp.domain.model.Store
-import com.example.buyerapp.domain.repository.GlobalKeyValueRepository
 import com.example.buyerapp.domain.repository.StoreRepository
 import kotlinx.coroutines.flow.FlowCollector
 import javax.inject.Inject
@@ -11,7 +10,7 @@ class SaveStoreUseCase @Inject constructor(private val storeRepository: StoreRep
     LocalUseCase<SaveStoreUseCase.Params, Unit>() {
 
     data class Params(
-        val id: Int,
+        val id: Long,
         val title: String,
         val descr: String,
         val color: String
