+++ app/src/main/java/com/example/buyerapp/presenter/promo/PromoScreen.kt
@@ -29,7 +29,7 @@ import kotlinx.coroutines.launch
 @Destination
 @Composable
 fun PromoScreen(
-    storeId: Int,
+    storeId: Long,
     viewModel: PromoViewModel = hiltViewModel(),
     navigator: NavigationProvider
 ) {
